refactor(grid): use tsx
fix #1409
